1:1대화 시작

+++ src/main/java/com/bbshop/bit/domain/OnetoOneVO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class OnetoOneVO {
 	
-	private int ONE_ONE_NUM;
+	private Long ONE_ONE_NUM;
 	private String ONE_CATEGORY;
 	private String ONE_TITLE;
 	private String ONE_CONTENTS;
